Eliminated pointless use of threads
StreamGobbler is an active buffering InputStream implementation. "active" because it has a thread that pumps the input stream, and "buffering", because it maintains an unbounded byte[] buffer that spools the data until it gets consumed.

This is both wasteful and harmful. Wasteful, because it uses two threads here (and one StreamGobbler is connected to a StreamCopyThread, making this a double-buffering. On systems that have a large # of slaves, this reduction in the total # of slaves is considerable.

This is harmful, because it eliminates the point of SSH window size control. If the master is under load, this will prevent the master from naturally throttle data that slaves are sending to it.