Eliminated the need for pumping stderr.
trilead-ssh2 build214-jenkins-3 has the ability to directly deliver
stdout/stderr data to OutputStream of our choice.

This version ships in Jenkins 1.521. This change allows us to take
advantage of this feature when we run with this version, and eliminates
the need for the stream pump thread.

On a large deployment, the thread saving would be considerable.