This plugin is confused about how the channel shutdown processing should happen.
Jenkins core calls afterDisconnect() from Channel.Listener.onClosed() (from the setChannel() method),
and this plugin installs its own Channel.Listener that has the onClosed() method.

The former gets called before the latter, and therefore some of the computation that was expected
to happen in the latter wasn't properly taking place. I'm consolidating both pieces of the shutdown logic
into a single function.

In this commit, the order of those two pieces are preserved, and one can already see that something isn't quite right here.

The next commit will clean up the afterDisconnect() function to make it do the proper thing