Clean up the channel loss handling.
Here, there are three layered connection-like things we need to think about.

1. SSH connection, which is a TCP/IP socket connection that multiplexes multiple "SSH channels" in it.
2. SSH channel, which is a logic sub-connection inside SSH connection that connects to the slave stdin/out and controls the process.
3. Remoting channel.

afterDisconnect() gets called when the #3 is lost, and hopefully when that happens #1 and/or #2 are still active. For example,
if the slave process has segfaulted, then SSH channel is still alive and it will bring the HotSpot crash dump of the JVM to the master
and the exit code. We want to capture that as it's a valuable diagnostic information.

We also want to know when the #1 is lost due to the underlying socket problem. This change doesn't address that, as the Connection class
needs to expose this information.