replace System.out.println with LOGGER
The System.out.println ends up on stdout of the controller jvm but
almost all other logout will end on stderr.
Also the formatting of the timestamp is hard coded and usually differs from
the normal logging output format.
By using the LOGGER we get well formatted log messages and all is in
stderr.